[front] refactor: create a re-usable layout for the ActionsPage
