+++ frontend/public/locales/fr/translation.json
@@ -714,6 +714,11 @@
     "whatRightsContributorsHaveOverTheirDataParagraph": "En accédant à la page de leur compte, les contributeurs peuvent télécharger les données transmises à la plateforme. Ils ont également la possibilité de demander la suppression de toutes les données personnelles les concernant que Tournesol conserve. Cela n'inclut pas les données que Tournesol a l'obligation de conserver pour des raisons administratives, légales ou à des fins de sécurité."
   },
   "clickToDownload": "Télécharger",
+  "actionsPage": {
+    "title": "Passez à l'action",
+    "whatCanYouDoToProtectSocietiesFromAIs": "Que pouvez-vous faire pour protéger les sociétés des conséquences des intelligences artificielles ?",
+    "whatCanYouDoToProtectYourselfAndYourRelatives": "Que pouvez-vous faire pour vous protéger et protéger vos proches ?"
+  },
   "myComparisonsPage": {
     "noComparisonYet": "Aucune comparaison",
     "listHasNbComparisons_one": "Vous avez déjà fait <1>{{comparisonCount}}</1> comparaison.",
